Change folder structure to enhance reusability
