+++ src/components/GameCard/GameCard.module.css
@@ -3,12 +3,6 @@
   border-radius: 15px;
   max-width: 400px;
   cursor: pointer;
-
-  --font-color: #fff;
-  --light-black: #27272a;
-  --lighter-black: #3f3f46;
-  --violet: #9333ea;
-  --gray: #71717a;
 }
 
 .card:hover {
