+++ src/components/Landing/landing.config.js
@@ -1,0 +1,40 @@
+const leftNavButtons = [
+  { name: 'Game Store', href: '/', iconSrc: './assets/icons/game-store.svg' },
+  {
+    name: 'Browse Store',
+    href: '/store',
+    iconSrc: './assets/icons/browse-store.svg',
+  },
+];
+
+const rightNavButtons = [
+  { name: 'trustinmagic', href: '', iconSrc: './assets/icons/github.svg' },
+  { name: 'Cart', href: '/cart', iconSrc: './assets/icons/cart.svg' },
+];
+
+const landingBottomNavButtons = [
+  { name: 'Browse', href: '/store', iconSrc: './assets/icons/browse.svg' },
+  { name: 'Play Dice', href: '', iconSrc: './assets/icons/dice.svg' },
+  { name: 'GitHub', href: '', iconSrc: './assets/icons/github2.svg' },
+  { name: 'LinkedIn', href: '', iconSrc: './assets/icons/linkedin.svg' },
+];
+
+const landingRightNavButtons = [
+  { name: 'Game Page', href: '', iconSrc: './assets/icons/game.svg' },
+  { name: '404 Page', href: '', iconSrc: './assets/icons/404.svg' },
+  { name: 'Commit Log', href: '', iconSrc: './assets/icons/git-commit.svg' },
+  {
+    name: 'Performance',
+    href: '',
+    iconSrc: './assets/icons/performance.svg',
+  },
+  { name: 'Technologies', href: '', iconSrc: './assets/icons/tech.svg' },
+  { name: 'Sources', href: '', iconSrc: './assets/icons/sources.svg' },
+];
+
+export {
+  leftNavButtons,
+  rightNavButtons,
+  landingBottomNavButtons,
+  landingRightNavButtons,
+};
