+++ src/components/Nav/Nav.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../styles/Nav.module.css';
+import styles from './Nav.module.css';
 
 const Nav = ({ leftButtons, rightButtons }) => {
   return (
