+++ src/components/Store/Store.module.css
@@ -2,7 +2,7 @@
   height: 100%;
   padding: 20px 40px;
   background-color: black;
-  color: white;  
+  color: var(--font-color)
 }
 
 .content {
