+++ src/components/StoreNav/StoreNav.module.css
@@ -2,8 +2,6 @@
   display: flex;
   flex-direction: column;
   gap: 20px;
-
-  --light-black: #27272a;
 }
 
 .filter,
